requirements.txt
Pillow
fontforge
